refactor: 컴폰너트 분리
