+++ src/features/member/components/MemberNicknameForm/style.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const InputAndButtonBox = styled.div`
+  display: flex;
+  align-items: start;
+  gap: 0.2rem;
+`;
